Read BGM

Read geometry and full topology from BGM files.

bgmfile(x, ...)

Arguments

x

path to a bgm file

...

ignored for now

Details

BGM is a file format used for the 'Box Geometry Model' in the Atlantis Ecosystem Model. This function reads everything from the .bgm file and returns it as linked tables.

See also

See helper functions to convert the bgm tables to `Spatial` objects, boxSpatial, faceSpatial, nodeSpatial, boundarySpatial, pointSpatial

Examples

library(bgmfiles)
bfile <- sample(bgmfiles(), 1L)
bgm <- bgmfile(bfile)

#> C:/inst/R/R/library/bgmfiles/extdata/bgm/ams71.bgm 
#> file declares 187 faces but contains data for 193 faces
#> 
#>  ... returning all 193 faces

library(tibble)
bgm

#> $vertices
#> # A tibble: 196 × 3
#>          x       y  .vx0
#>      <dbl>   <dbl> <int>
#> 1  1529167 2084608     1
#> 2  1535624 2033062     2
#> 3  1538150 2013366     3
#> 4  1564898 2089063     4
#> 5  1603981 2072002     5
#> 6  1617781 2051053     6
#> 7  1663541 2168770     7
#> 8  1694033 2155419     8
#> 9  1707112 2220880     9
#> 10 1722830 2137410    10
#> # ... with 186 more rows
#> 
#> $facesXverts
#> # A tibble: 386 × 3
#>     .fx0   .p0  .vx0
#>    <dbl> <int> <int>
#> 1      0     1     4
#> 2      0     2     5
#> 3      1     1   149
#> 4      1     2   155
#> 5      2     1   155
#> 6      2     2   159
#> 7      3     1   159
#> 8      3     2   172
#> 9      4     1   172
#> 10     4     2   179
#> # ... with 376 more rows
#> 
#> $faces
#> # A tibble: 193 × 7
#>         cosine       sine  left right    length  .fx0 label
#>          <dbl>      <dbl> <int> <int>     <dbl> <dbl> <chr>
#> 1  -0.91648273 0.40007425     1     0  42644.89     0 face0
#> 2  -0.99326818 0.11583746    20    22  31404.53     1 face1
#> 3  -0.98651191 0.16368948    23    24  13055.24     2 face2
#> 4  -0.99883874 0.04817849    47    48 123323.96     3 face3
#> 5  -0.98408116 0.17771967    49    50 115328.51     4 face4
#> 6  -0.99701854 0.07716237    54    49 308798.85     5 face5
#> 7  -0.97551285 0.21994246    49    56  96618.49     6 face6
#> 8   0.04213626 0.99911187    24    20 158232.50     7 face7
#> 9   0.06564522 0.99784303    47    24 159084.28     8 face8
#> 10  0.66566271 0.74625274    24    20  37031.67     9 face9
#> # ... with 183 more rows
#> 
#> $facesXboxes
#> # A tibble: 385 × 3
#>     .bx0 iface  ibox
#>    <dbl> <dbl> <dbl>
#> 1      0     0     1
#> 2      1     0     0
#> 3      1    62     2
#> 4      1    70    34
#> 5      1    81    34
#> 6      2    42     3
#> 7      2    43    32
#> 8      2    44    33
#> 9      2    61    33
#> 10     2    62     1
#> # ... with 375 more rows
#> 
#> $boxesXverts
#> # A tibble: 542 × 2
#>     .bx0  .vx0
#>    <dbl> <int>
#> 1      0     1
#> 2      0     4
#> 3      0     5
#> 4      0     2
#> 5      0     1
#> 6      1     9
#> 7      1    13
#> 8      1    14
#> 9      1     8
#> 10     1     5
#> # ... with 532 more rows
#> 
#> $boxes
#> # A tibble: 71 × 9
#>           label nconn  botz         area vertmix horizmix insideX insideY  .bx0
#>           <chr> <int> <int>        <dbl>   <dbl>    <int>   <dbl>   <dbl> <dbl>
#> 1  BoundaryBox0     1  -200   2279336755   1e-06        1 1560270 2063300     0
#> 2          box1     4  -200  33076861706   1e-06        1 1848738 2182818     1
#> 3          box2     5  -200 106000000000   1e-06        1 2461586 2395124     2
#> 4          box3     4  -200  52960144011   1e-06        1 2846262 2423700     3
#> 5          box4     7  -200  57204362952   1e-06        1 3043747 2204414     4
#> 6          box5     1   -20  20794488111   1e-06        1 3271770 2274790     5
#> 7          box6     7  -200  32236659536   1e-06        1 3367598 2017187     6
#> 8          box7     3  -200   9839460116   5e-05        1 3571038 1860595     7
#> 9          box8     9  -200  30312646954   1e-06        1 3828736 1629875     8
#> 10         box9     7   -50  71053969415   1e-06        1 4036585 1604810     9
#> # ... with 61 more rows
#> 
#> $boundaryvertices
#> # A tibble: 67 × 3
#>          x       y bndvert
#>      <dbl>   <dbl>   <int>
#> 1  4952047 3145085       1
#> 2  4964926 3142948       2
#> 3  5088107 3137007       3
#> 4  5201600 3116510       4
#> 5  5509478 3092683       5
#> 6  5603730 3071432       6
#> 7  5545009 2853190       7
#> 8  5643228 2683493       8
#> 9  5487038 2568810       9
#> 10 5373099 2279576      10
#> # ... with 57 more rows
#> 
#> $extra
#> $extra$maxwcbotz
#> [1] "-1800"
#> 
#> $extra$nbox
#> [1] "71"
#> 
#> $extra$nface
#> [1] "187"
#> 
#> $extra$projection
#> [1] "+proj=aea +lat_1=-18 +lat_2=-36 +lat_0=0 +lon_0=134 +x_0=3000000 +y_0=6000000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
#> 
#>