Utilities for BGM files for Atlantis

Tools for handling network data for Atlantis from box-geometry model (BGM) files

rbgm features

read .bgm files and faithfully store all information so it can be round-tripped

conversion from .bgm forms to Spatial classes (lines and polygons)

(not yet implemented: write to .bgm)

I. Import

bgmfile	read directly from a .bgm file

II. Conversion

boxSpatial	convert boxes to a SpatialPolygonsDataFrame

faceSpatial	convert faces to a SpatialLinesDataFrame

boundarySpatial	convert boundary to a single-row SpatialPolygonsDataFrame

nodeSpatial	obtain all vertices as points

pointSpatial	obtain all instances of vertices as points

III. Miscellaneous

build_dz	Build Atlantis dz Values